Extract QtPanelBuilder into graphical/qt_panels.py
qt_gui.py split, final cut (item 1 for Qt). The recursive panel/control
widget builders move into a stateless QtPanelBuilder that keeps all mutable
registries (widget_list, labels, timer) on the window, exactly as before.
MainWindow retains thin _build_panel_widget/_create_control_widget
delegates with unchanged signatures for DetachedPanelWindow and callers.
qt_gui.py: 976 lines before the split, ~500 after.
