+++ src/interactive_pipe/graphical/qt_gui.py
@@ -45,6 +45,7 @@ from interactive_pipe.graphical.qt_image import (
     update_curve_cell,
     update_table_cell,
 )
+from interactive_pipe.graphical.qt_panels import QtPanelBuilder
 
 # Re-exported for backward compatibility (they lived in this module until 0.9.x)
 from interactive_pipe.graphical.qt_widgets import CollapsibleBox, DetachedPanelWindow  # noqa: F401
@@ -333,167 +334,19 @@ class MainWindow(QtWidgetBase, InteractivePipeWindow):  # type: ignore[reportGen
             self.main_gui.on_press(mapped_str, refresh_func=self.refresh)  # type: ignore[reportOptionalMemberAccess]
 
     def _build_element_widget(self, element, control_factory: ControlFactory) -> Optional[QWidget]:
-        """Build widget for an element (Panel or Control).
-
-        Args:
-            element: Can be a Panel or Control instance
-            control_factory: Factory for creating control widgets
-
-        Returns:
-            QWidget or None if element should be skipped
-        """
-        if isinstance(element, Panel):
-            return self._build_panel_widget(element, control_factory)
-        elif isinstance(element, Control):
-            return self._create_control_widget(element, control_factory)
-        return None
+        """Build widget for an element (Panel or Control). Delegates to QtPanelBuilder."""
+        return QtPanelBuilder(self, control_factory).build_element_widget(element)
 
     def _build_panel_widget(self, panel: Panel, control_factory: ControlFactory) -> QWidget:
-        """Recursively build Qt widget for a Panel.
+        """Recursively build Qt widget for a Panel. Delegates to QtPanelBuilder.
 
-        Args:
-            panel: The Panel to build
-            control_factory: Factory for creating control widgets
-
-        Returns:
-            QWidget containing the panel's content (CollapsibleBox or QGroupBox)
+        Keep this signature stable: DetachedPanelWindow calls it.
         """
-        # Determine layout based on elements structure
-        if panel.elements and isinstance(panel.elements[0], list):
-            # Grid layout: list of lists
-            layout = QGridLayout()
-            for row_idx, row in enumerate(panel.elements):  # type: ignore[reportArgumentType]
-                if isinstance(row, list):
-                    for col_idx, element in enumerate(row):
-                        widget = self._build_element_widget(element, control_factory)
-                        if widget is not None:
-                            layout.addWidget(widget, row_idx, col_idx)
-        elif panel.elements:
-            # Vertical layout: flat list
-            layout = QVBoxLayout()
-            layout.setSpacing(8)  # Nice spacing between controls
-            for element in panel.elements:
-                widget = self._build_element_widget(element, control_factory)
-                if widget is not None:
-                    layout.addWidget(widget)
-        else:
-            # No child elements, just a simple vertical layout for controls
-            layout = QVBoxLayout()
-            layout.setSpacing(8)  # Nice spacing between controls
-
-        # Add controls assigned directly to this panel
-        for ctrl in panel._controls:
-            widget = self._create_control_widget(ctrl, control_factory)
-            if widget is not None:
-                layout.addWidget(widget)
-
-        # Add stretch at the end to push all controls to the top
-        # This creates a beautiful, balanced layout with controls at the top
-        if isinstance(layout, QVBoxLayout):
-            layout.addStretch()  # type: ignore[reportAttributeAccessIssue]
-
-        # Create the panel widget (collapsible or regular)
-        if panel.collapsible:
-            # Use modern CollapsibleBox
-            panel_widget = CollapsibleBox(title=panel.name or "", collapsed=panel.collapsed)
-            panel_widget.set_layout(layout)
-        else:
-            # Use regular QGroupBox
-            panel_widget = QGroupBox(panel.name or "")
-            # Set nice margins for balanced appearance
-            if isinstance(layout, QVBoxLayout):
-                layout.setContentsMargins(12, 12, 12, 12)
-            panel_widget.setLayout(layout)
-
-        return panel_widget
+        return QtPanelBuilder(self, control_factory).build_panel_widget(panel)
 
     def _create_control_widget(self, ctrl: Control, control_factory: ControlFactory) -> Optional[QWidget]:
-        """Helper method to create a single control widget with labels.
-        Returns the row container widget or None if control should be skipped."""
-        slider_name = ctrl.name
-
-        if isinstance(ctrl, KeyboardControl):
-            if self.main_gui is not None:
-                self.main_gui.bind_keyboard_slider(ctrl, self.key_update_parameter)  # type: ignore[reportOptionalMemberAccess]
-            return None
-        elif isinstance(ctrl, TimeControl):
-            self.timer = QTimer(self)
-
-            def suspend_resume_timer(suspend: bool):
-                if suspend:
-                    logging.debug("Suspend")
-                    self.timer.stop()
-                else:
-                    logging.debug("Resume")
-                    self.timer.start()
-
-            if self.main_gui is not None:
-                self.main_gui.suspend_resume_timer = suspend_resume_timer  # type: ignore[reportOptionalMemberAccess]
-                plugged_func = self.main_gui.plug_timer_control(ctrl, self.update_parameter, suspend_resume_timer)  # type: ignore[reportOptionalMemberAccess]
-            else:
-
-                def plugged_func():
-                    pass
-
-            self.timer.timeout.connect(plugged_func)
-            self.timer.start(ctrl.update_interval_ms)
-            return None
-        elif isinstance(ctrl, Control):
-            slider_instance = control_factory.create_control(ctrl, self.update_parameter)
-            # Skip controls that return None (e.g., single-value controls)
-            if slider_instance is None:
-                return None
-            if ctrl._type is str and ctrl.icons is not None:
-                if ctrl.filter_to_connect is not None:
-                    ctrl.filter_to_connect.cache = False  # type: ignore[reportOptionalMemberAccess]
-                    ctrl.filter_to_connect.cache_mem = None  # type: ignore[reportOptionalMemberAccess]
-                # Disable cache for dropdown menu with icons!
-                # Allows clicking on the same item multiple times
-            slider_or_layout = slider_instance.create()
-            self.widget_list[slider_name] = slider_instance
-
-            slider_layout = QHBoxLayout()
-            slider_layout.setSpacing(8)  # Nice spacing between label, slider, and value
-
-            if isinstance(slider_or_layout, QWidget):
-                # Balanced label width that works well in both panels and bottom area
-                label_fixed_width = 150
-                label = QLabel("", self)
-                label.setMinimumWidth(label_fixed_width)
-                self.name_label[slider_name] = label
-                slider_layout.addWidget(self.name_label[slider_name])
-            if isinstance(slider_or_layout, QWidget):
-                # If it's a QWidget, add it directly to the layout
-                # Set minimum width to prevent slider from shrinking in narrow panels
-                slider_or_layout.setMinimumWidth(150)
-                # Add with stretch factor 1 so slider expands to fill available space
-                slider_layout.addWidget(slider_or_layout, 1)
-            elif isinstance(slider_or_layout, QHBoxLayout):
-                slider_or_layout.setContentsMargins(0, 0, 0, 0)
-                # If it's a QHBoxLayout, embed it in a QWidget first
-                container_widget = QWidget()
-                container_widget.setLayout(slider_or_layout)
-                # Set minimum width to prevent shrinking in narrow panels
-                container_widget.setMinimumWidth(150)
-                slider_layout.addWidget(container_widget, 1)
-            else:
-                logging.warning(f"Unhandled type for slider: {type(slider_or_layout)}")
-                return None
-            if isinstance(slider_or_layout, QWidget):
-                result_fixed_width = 90
-                label = QLabel("", self)
-                label.setMinimumWidth(result_fixed_width)
-                self.result_label[slider_name] = label
-                slider_layout.addWidget(self.result_label[slider_name])
-
-            # Create a container widget for the entire row
-            row_container_widget = QWidget()
-            row_container_widget.setLayout(slider_layout)
-            row_container_widget.setContentsMargins(0, 0, 0, 0)
-
-            self.update_label(slider_name)
-            return row_container_widget
-        return None
+        """Create a single control widget row. Delegates to QtPanelBuilder."""
+        return QtPanelBuilder(self, control_factory).create_control_widget(ctrl)
 
     def init_sliders(self, controls: List[Control]):
         self.ctrl = {}
